Create disclaimer.md
This can be used on the website:  (1) should be put on the header or footer, and should link to (2).  These are NOT full terms of service/use, and are NOT a privacy policy (although it can be integrated into ToS).

However, I think it's valuable to have them as a standalone disclaimer in the repo, as well.